Move UserPage to users/UsersRootPage
