+++ kolibri/plugins/facility/assets/src/views/users/UsersRootPage/DeleteUserModal.vue
